Rewrite depth-first iterator.
Does not access the pool indices anymore, but is based on instead of
PoolObjectCollection classes. Way more elegant.

The problem now is that the algo needs to access the vertex pool to
build the pool collection objects and the two pools of a graph are not
accessible.

By the way the graph interface is expected not to have any pool, since
it is supposed to be general. Let's see how we deal with that with other
algos.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>