Rewrite DFI according to Tobias collection framework.
I am not sure it works. There is an interesting test failure in the JUnit test
provided here.  I simply could not keep a refrence to the previously iterated
node as the reference is changed by the DFI, but the actual object is the same.
This is somewhat expected, but I want to be sure I am in synch with Tobias.

Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>