added CollectionUtils.safeIterator() that wraps (if necessary) an Iterator such that objects returned by next() stay valid over successive calls. Using this fixes DFITest.
However, it would be nice to have a mechanism to be able to write such code
in a way that works for refs and normal objects and doesn't need to allocate
many ref objects. For ref objects, the refTo() method can be used to do
assignment such as "previous = vertex" in DFITest. We need something that
works also for "normal" objects.