Rework the tree outputter.
Right now it compiles, but it still crashes for graphs that are not trees.
We want it to silently fail for such graphs, not crash.
Heaviliy NOT optimized. The path chosen to deal with maps where vertices
are key was the most simple one, not the most efficient one.
To review.

Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>