A Graph can return iterators that iterate over all edges or vertices.
If we receive a graph instance only, we cannot possible navigate through its
content without at least a vertex or an edge. We need some kind of method that
gives access to the graph content.

Using the Iterator interface is probably safer. It is unwise to return a
collection of the graph content. For instance if in the future we ever
choose to override an implementation with a virtual structure, having to
override a method that returns or exposes the whole content of the graph
is going to be very difficult. Whereas returning a specialized Iterator
is much simpler.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>