A tentative, default connected components analysis for dynamic graph.
This implementation is devoid of any subtleties that exists for instance
in [1]. The reason behind this choice is that I presume that a large
majority of the tracks in TrackMate will be tree-like. So any edge
deletion in such a track is sure to be a tree-edge, and to generate two
connected components. So we rely on brute force (iterating over one side
of the cut cc), saving the memory needed to maintain a spnning-forest of
the ccs.

TODO: tests tests tests

[1] Jacob Holm, Kristian de Lichtenberg, and Mikkel Thorup. 2001.
Poly-logarithmic deterministic fully-dynamic algorithms for
connectivity, minimum spanning tree, 2-edge, and biconnectivity. J. ACM
48, 4 (July 2001), 723-760. DOI=10.1145/502090.502095
http://doi.acm.org/10.1145/502090.502095

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>