Add maps from vertex or edge to plain int.
Useful to associate a vetex or a map to a level in standard graph algorithms.
Very specialized, but useful for graphs.

Its implementations exist to take advantage of the compact data
storage offered on one side by the Trove library for standard object, and of a
derivative of a Trove class to deal specifically and advantageously with
PoolObjects. Here, we therefore decorate the Trove mother interface with extra
methods that accept an existing Ref object to control garbage collection.

Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>