A SelectionHandler that implements the previous TrackScheme selection behavior:
Single-click on a vertex or an edge to replace the selection by this vertex or edge.
Single-click elsewhere to clear the selection.
Click and drag to draw a box around selection.
Shift - Single-click to toggle selection with the target vertex or edge.
Shift - Click and drag to add a box content to the current selection.

Signed-off-by: Jean-Yves Tinevez <tinevez@pasteur.fr>