Rework the wasy we listen to graph changes.
There is now two level of events fired:

- GraphListeners will be notified on the finest granularity possible:
	- after a vertex or an edge have been added;
	- before a vertex or an edge are removed,
they are immediately notfied via void methods.
	- they are notified that the user called beingUpdate(),
	  endUpdate(), etc. the same way. They can choose to ignore
this.

- GraphChangeEventListaners have to subscribe to the
GraphChangeEventAggregator. This guy is a GraphListener, and it is in
charge of aggregating fine granularity events into a GraphChangeEvent
object that everything that happened between beginUpdate() and
endUpdate().
  It is well suited e.g. to repaint something after a mass delete from
the user, or to write an UNDO mechanism.
  If changes are made outside of a beginUpdate() / endUpdate() block,
events are fired immediately.
  The two methods pauseUpdate() and resumeUpdate() are use to make it
listen or not.

Signed-off-by: Jean-Yves Tinevez <tinevez@pasteur.fr>