Move concrete Spot stuff in two different packages.
The problem is that now we cannot use the default visibility
anymore. So some of the guts of the graph is now public visible.
But let's say that this is the choice: we want possibly to
choose how Spot are represented and it makes sense to put
everything that relates to this specificity into one single
package.

Let's see how it turns out for the interaction with the bdv.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>