Move the highlight logic to its handler.
Before thie commit, the logic that sets and unsets the highlight in
TrackScheme was dealt with in the AbstractTrackSchemeOverlay class. Now
this class only serves what TrackScheme vertex and edge id are close to
a mouse location (their id actually), and it is the responsibility of
the MouseOverListener to maintain the highlight object.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>