Use the selection navigator in TrackScheme.
Using the arrow key will navigate the currently selected
vertex in the 4 directions. With shift, the navigated vertex
is added to the selection.

The usability of this has still to be reviewed, though, because
we do not center on the navigated vertex. This will be addressed
when we deal with centering.

Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>