WIP: Try to derive a framerowk for shared "navigation" between views.
The idea is for views (a BDV window or a TrackScheme window) to register
themselves in locks. If two views have the same lock toggled on, a navigation
event in one of the view will caused other views to be centered on the
same spot.

I stole the idea of the locks to Icy.

This is cool, but the core of the work is missing. We need to define
a navigation event, and must be able to trace what view generated it.
This is not here right now. The best suspect is to play a bit with
the SelectionNavigator to see if it is applicable.

Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>