Shared navigation across views, only for TrackScheme yet.
Logic:
- A NavigationHandler manages a list registered NavigationListeners,
that can belong to 0, 1 or several groups.
- When a selection event is triggered on a view, we check if there is
only one vertex selected.
- If yes, an event is triggered from this view, and forwarded to all
the navigation listeners that belong to the same group(s) that of
the triggering view.
- Receiving views are then centered on the designated vertex.

On TrackScheme, a panel with 5 locks determine what groups this view
belongs to.

What does not work yet:
- When centering, the translation is not animated.
- Maybe there is no need to move the panel if the target vertex is
visible on the current view.
- We need an actual instance of the TransformEventHandler, which is
uneasy to get now.
- Selection box events trigger false single-vertex-selected events.

Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>