The HighlightNavigator can trigger shared navigation.
I made the following fields of TrackSchemePanel with the package visibility:
- layout
- display

Arrow key navigation happens from the highlighter vertex. If there is no
highlighted vertex, we take the one closest to the panel center. If one
lock is activated in the lock panel, the view will be updated to follow
the highlight vertex (as long as it is highligted with keys). It is put
in the middle of the panel.

Using the highlight instead of the selection is a far better choice.
Now we can use this to edit selection with the keyboard TODO.

Now when a lock is activate in the lock panel,
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>