The navigation handler and listener are type with vertex type.
Instead of passing the model vertex id, we directly pass the vertex
object itself, and type NavigationListener and NavigationHandler.
The problem is that we now have to pass the model graph and idmap
to retrieve the model vertex to fire from the TrackSchemeVertex.

Let's see this when we address the same problem for selection.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>