Rework the navigation logic for TrackScheme.
There is now a FocusModel, which is managed in TrackScheme (only there
for now) as the HighlightModel. The focus vertex is used as a cue for
navigation. Keyboard arrow keys move the focus around.
But this is independant of the highlight, which is now mouse only.

While at it, implements double-click on a vertex to navigate to it in
TrackScheme.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>