Pause and resume selection handler to avoid firing 100000 events.
This is what was causing the selection animation bug. When we select a large
number of vertices, or select a vertex while there is already a selection,
several if not many events are fired. By putting a compound selection change
between pause and resume, we only fire one event.
This fixes the selection animation bug.

Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>