Change the way grouping is handled.
The concept of grouping was separated from of the individual model UIs.
For example, TrackSchemeNavigation does not need to know a
NavigationGroupHandler anymore. Each individual model UI talks to its
own separate NavigationHandler, which does not reflect any information
about groups. In fact, it would be easy (but not required currently) to
make NavigationHandler an interface which then could be implemented to
support different kinds of grouping or no grouping at all. The point is,
the TrackScheme UI doesn't care. It sends a single notify event to its
NavigationHandler, which knows whether to distribute events to other group
members or not.

Moreover, grouping has been split from the specifics of NavigationHandler.
There might be other stuff that needs grouping (e.g. Focus) and we now
support that easily. The way it works is as follows:
1.) Create a GroupManager.
2.) For each entity that has group membership (i.e. everything that has a
    LocksPanel) create a GroupHandle via GroupManager.createGroupHandle().
    If required, a LocksPanel for the entity can be created with
        new GroupLocksPanel(groupHandle);
3.) Attach stuff to the GroupHandle. For example, NavigationHandler is
    constructed with a GroupHandle and the constructor calls
        GroupHandle.add(this);
    The GroupHandle keeps a registry of attached things by Type.
    It can talk to the GroupManager to find things of the same Type in
    other GroupHandles that share active locks with this one.
4.) A NavigationHandler instance can then ask its GroupHandle for other
    NavigationHandlers that share (via their GroupHandle) active locks.
    The loop to distribute notify these linked NavigationHandlers is now
    simply this:
      for (NavigationHandler<V> h : group.allInSharedGroups(this))
        for (NavigationListener<V> l : h.listeners)
          l.navigateToVertex(vertex);

The GroupLocksPanel doesn't need to know about NavigationHandlers.
The NavigationHandler doesn't need to know about the mechanics of grouping.
The clients of NavigationHandler don't need to know about grouping at all.

If we want to apply grouping to Focus, we can simply
      GroupHandle.add(this);
and
      for (FocusHandler<V> h : group.allInSharedGroups(this))
without implementing any additional logic.