User ArrayList instead of HashSet to store listeners.
Following Tobias input:
"The handlers often have Sets of listeners.
I want to swap all of the implementations that currently use HashSet to using
ArrayList (but acting like a set, i.e., no duplicates). It is already like that
in net.trackmate.revised.ui.selection.HighLightModel.  The reason is that for
dispatching events we iterate over the listeners and iteration is much faster
for ArrayList. It's kind of obvious, but I also found this article that
confirms it:
http://www.javacodegeeks.com/2010/08/java-best-practices-vector-arraylist.html
(Test case #3 Iterators). We very seldom add/remove listeners, but very
often iterate over them, so..."

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>