fix javadoc errors (java 8). Involved manual fixes and: 1) find src -name *.java -exec sed -i '' -e 's/\(@author[^>]*\)<\([^>]*\)>/\1\&lt;\2\&gt;/g' {} \; to fix author email addresses with "<" and ">"
2)
find src -name *.java -exec sed -i '' -e 's/<code>/{@code /g' {} \;
find src -name *.java -exec sed -i '' -e 's/<\/code>/}/g' {} \;
to replace "<code>...</code>" by "{@code ...}".
Apparently, for the latter "<" and ">" are ok, but not for the former.
For example, "<code>A -> B</code>" is an error, but "{@code A -> B}" is not.