Make the changes in covariance undoable.
This is a bit hacky compared to existing design:
The covariance field can be set in the Spot concrete class. But the model
we use in the mamut package is composed with the undo recorder that is
generic for AbstractSpot.

So I recreated the Undo system, making it specifics to Spot with
covariance, and change the Model to use this system.