Action to create a spot that links to the spot we clicked into.
In the BDV, it is bound to 'A', like for creating spot, except that it
does nothing if we do not click inside a spot, which makes it possible
to share the same key mapping than for the AddSpot action.

I put it in a separate class, because
- it needs to access the ViewerPanel to change the current timepoint;
- other actions of the same kind might need to use an extra overlay in
the future.