A BDV action to toggle links by drag behaviour.
Press L inside a spot.
The viewer is moved to the next panel and the source spot is displayed
as a ghost. As you drag, a line is drawn to the mouse.
Release inside a target spot to create a link between source and target.

If there is already a link between the two, the link should be removed
but I do not know how to do that through the OverlayGraph, as it is
a ReadOnlyGraph.