In BDV, the highlight picks an edge first (was a vertex before).
The reason is that in practice, links are much shorter than the
typical vertex radius. So all the edges that are inside a vertex
could not be picked up by the highlight.
With this simple change they can. It is not too penalizing, because
the way we compute distance to edge ensures that there is always
space in the vertex where it can be picked up.