Push the notifyGraphChanged methods and co to Listenable graph interface, and make them public.
So that classes that modified the graph can notify that they did.

@tpietsch, is thit ok?
We want to build actions that modify the graph, and have our views
be aware of the modifications. A more classical approach would be
to have the views register as a listener to graph changes, but for
other actions, this approach (calling motifyGraphChanged) is used.