Add the possibility to add overlays to the TrackScheme overlay.
The idea is that some behaviours might want to paint something
explenatory on the TrackScheme panel. To be nice, you want to
paint it over the background, but below the edges, vertices, etc.
To do so, we add a list of overlay to the TrackScheme of overlay
and defer painting to them.

Another change required: if we have to use this overlay, we need
to expose it. This last point might change with the future
overhaul in TrackScheme instantiation.