A non working edge editor for TrackScheme.
FIXME: Does not work in practice for MaMuT, because Spots
and Links need to be init() after being added to the
model graph.

In TrackScheme, we do not have access to the concrete edge
and vertex class. There are model properties, but they are not
meant for writing stuff.

Trying to add 2 links with this method will generate an
exception.