Finder-like behaviours for selection/focus in TrackScheme.
Two versions of TrackScheme Navigators are now available
as specified by Tobias:

1. 'Finder-like' behaviour (currently used).

Selection is tied to the focus. When moving the focus with
arrow keys, the selection moves with the focus. When clicking
a vertex it is focused and selected, The focused vertex is
always selected. Focus still exists independent of selection:
multiple vertices can be selected, but only one of them can
have the focus. When extending the selection with shift+arrow
keys, the vertex to which the focus moves should be selected.
When box selection is drawn, the selected vertex closest to
the position where the drag ended should receive the focus.

2. 'Midnight-commander-like' behaviour.

Selection is independent of focus. Moving the focus with arrow
keys doesn't alter selection. Space key toggles selection of
focused vertex. When extending the selection with shift+arrow
keys, the selection of the currently focused vertex is toggled,
then the focus is moved.