Common selection behaviours to delete a selection.
An AbstractBehaviours class, common to TS and BDV that operates
on a Selection only. Right now only one behaviour is implemented:
delete selection (shift + DELETE).

Edges must be removed first or the JVM crashes.
I don't know if removal is efficient as it is. Should we pause
listeners for batch removal?