WIP: Share list of BDV styles across BDVs.
To do this we rely on a Combo-box model that has a shared
list of objects. The combo boxes update the list they show
automatically.
Still to do:

- Centralize loading/saving the list. As of now we reload the list
everytime we instantitate a BDV and prune duplicates. This is
ineffective.

- The BDVs are not automatically notified when a syle is changed in
anotehr BDV. Right now: the last BDV to edit a style sets it,
regardless of what the other do. Also, if a style is deleted in a
BDV and used in another, the later is not automatically changed.
Maybe this is good enough?