Trim FeatureComputer and co.
- FeatureComputers are stored directly in the ComputerService.
They do not declare what they compute and on what objects what they
compute is defined for. But this is fine for the rest of the
application.

- So I could get rid of most of the methods based on String, and on
methods specific to edges or vertices.

- However I kept the marker interface SpotFeatureComputer and
LinkFeatureComputer. They are specific to MaMuT and used for
SciJava discovery. I was affraid that if I collect all the instances
of FeatureComputer, I would also import in MaMuT FeatureComputers
that are defined for something else than MaMuT.