Revised how user-oriented feature are saved.
- Feature de/serialization is triggered from the `FeatureModel`, when the
model is being serialized itself.

- Features are de/serialized from/to a subfolder of the project folder.
With the `DefaultFeatureModel,` this subfolder is 'Features'.

- Each feature is saved to an individual file, named with the key of the
feature + '.raw'.

- Specialized classes, derived from the interface `FeatureSerializer`,
are in charge of the de/serialization of individual features. There is
*one* specific `FeatureSerializer` for each feature. This is necessary
since we do not know what is the type of values used by a feature.

- Concrete `FeatureSerializer` implementations are provided by the
`FeatureComputer`s. Since these classes (normally written by 3rd party
users) are the ones that know how to generate a feature, they also know
how to de/serialize them.

- So the model needs to know what are the `FeatureComputer`s when saving
and loading. Since they are provided by SciJava plugin discovery mechanism
(within the FeautureComputerService), the model needs to access a SciJava
`Context`. Now each `Model` instance has its own context (lightweight,
only instantiated with the `MamutFeatureComputerService`). The context
instance has then been removed from the `MainWindow` class, and the later
uses the context of the model it manages.

Dangers:

- Since the feature files are separated from the model file, there might
be a situation where the user loads a model that has been modified after
the feature file has been saved (e.g. the user saves a subset of features,
but there remains a feature file from a previous save that is not
overwritten). I do not know what will happen. To test. Possible solution:
wipe the whole 'Features' folder prior to saving a project.