Add SettingsPage for selecting between and editing named profiles.
This handles the following scenario:
There are multiple "profiles" (e.g., keymaps, trackscheme styles, ...)
to select from, and its possible to edit each individual profile.
Profiles can be renames, duplicated, deleted. Some profiles are "built-in"
and cannot be modified (will be duplicated when edited). Selecting a
different profile and editing the current profile both constitute
modifications that are reported to ModificationListeners.