Add a flag to TrackSchemePanel to signal for TrackScheme entites attributes changed.
This flag is meant to signal that some of the attributes of the TrackScheme
entities to paint have changed, without the actual layout of these entities
changing.

This triggers a single `cropAndScale` call on the layout, recopying the
attributes of the entities to the Screen entities without animation.

This qualifies for e.g. use cases where the color of vertices is changed
(to be implemented) or when the label of a vertex is changed (the purpose
of this branch).