File format v0.2 proposal.
- Mastodon projects are saved in a folder. The project folder is the
project name (user convention) with a '.mastodon' suffix.

- Model files are always saved directly within this folder, and always
have the name 'model.raw'.

- Project files are always saved directly within this folder, and always
have the name 'project.xml'. They contain the relative path to the BDV
xml file, and the relative path to the model file (in case of people
still wanted to access a model file with another name, another location).