Fix hide-show toolbar action resizing the view.
Prior to this commit, hiding/showing the toolbar of Mastodon views
(e.g. by pressing 'T' in TrackScheme) caused the view to be resized
to its default size.

This commit fixes this bug by setting the preferred size of the view
frame before re-packing it before deletion/insertion of the toolbar.