Deal with invalid feature color mode.
The declared feature color modes might rely on features that are not
computed in the feature model yet. In that case, the UI for configuring
feature color modes shows a warning message that states what features are
missing. Also, the TrackScheme coloring menu shows invalid color modes,
but they are disabled.