Rework the feature and feature calculation frameworks.
- Feature is now an interface. Its role is limited to expose read-only
values for a collection of objects of the same target class. The
values can be anything, but - as before - can be projected over a list
of scalar double values.

- A feature knows how to serialize itself to a file. This is not used yet.

- There are a couple of abstract classes derived from this interface,
made to simplify generating them. There are one for scalar double and
one for int features, that wraps a PropertyMap. There is also one for
double array values, that wraps a *list* of property maps.
  Regarding this point: We could have used a compact storage, a la RefPool,
where the array for a single object is stack in a large array.
  Here I favored a list of scale property maps, because I reckon that we
will need performance when iterating over a single element of this
array over the whole collection of objects. For instance, if we generate
a coloring  based on the mean intensity value of all spots, we do not need
to access the std value for a spot, which would be stored next to the
mean in a compact storage.
This is not true when generating a table of all the feature values, but
this needs not to be super responsive.

- Mastodon app (Mamut) feature computers can generate such features.
There are now extra abstract class to facilitate writing e.g. scalar
feature computers.

- Feature computers know how to deserialize their feature from a file.
This is not used yet.