Track features.
Track features have nothing specific made for them.
They are actually spot or link feature where each spot or link has the
feature value of the track it belongs to.
Each track feature can have a the '"Spot track ID' that returns what
is the track id of a spot, and use it to store the track feature
value in the spots of the said track.