Change FeatureProjection: int marker interface and return units.
FeatureProjection can now returns its  physical units as string.
This aims at supporting units in Mastodon for feature values.
  There are facilities to generate a unit string from the dimension
of a quantity and spatial and temporal units.
  Also, there is a marker interface for FeatureProjection that returns
integer values. This is aimed at being used in user interfaces such as
tables, where we want to display the integer values nicely.
  This breaks the current MaMut FeatureComputers, but not for long.