Change the MaMut-specific feature computers.
We want to write feature values that depend on image data (for instance
something like spot mean intensity). For this the feature computer must
access the BDV data. Since FeatureComputers are SciJava plugin, they must
receive it after instantiation from a method that will be called by the
feature computer service.

- So now there is a MamutFeatureComputer abstract class that does this.

- And it has facilities to generate unit strings for quantities in
FeatureProjections, using the spatial units stored in the first Source.

- And we had utility abstract classes that cover common use cases (int
and double, scalar or array feature computers) with deserialization.