Spot feature computer that computes gaussian mean intensity.
This feature computer exploits the BDV data this it receives
to iterate over the pixels inside all spots. To compute mean
intensity we generate a gaussian kernel with a size limited
by its smallest radius (only account for pixels inside the
ellipsoid).

Also this feature computer is the first with a proper config panel.
It allows for specifying over what sources to compute the mean
intensity. Shutting down some sources might be important because
the computation can be very long for some large models.