Classes to support incremental feature computation.
A GraphUpdateStack instance will record graph changes between
certain 'commits'.

After features have been computed, a commit is made to the stack by
the feature computer service. The commit stores the list of feature
keys that have been computed. The stack can store a limited number
of commits (10).

When needed, a FeatureComputer can call the #changesFor( key ) method,
which will return an GraphUpdate object containing all the changes that
happened in the graph since the last time the feature was computed.
This object can be used to make an incremental computation of the feature.

The GraphUpdate object is produced by concatenating all the GraphUpdates
over all commits until the commit with the feature key is found.

If a GraphUpdate object cannot be produced (is null, because e.g. the
feature was never computed, or because the last commit with this
feature was 'forgotten' by the stack), this signals that a full
computation must be made.

There are JUnit tests for this mechanism that uses a method from
HamCrest I had to copy to avoid conflict with maven enforcer.