Make FeatureComputerService cancelable and listenable.
- FeatureComputerService is now an interface.
- It's default implementation is still generic.
- It is cancelable, and will forward cancel events to the current
feature computer if it cancelable.
- It can accept a PropertyChangeListener that will be provided to
feature computer as a dependency, if they want to report progress.
- It can return a feature computer for a feature spec if it knows
about it.

The last part is important, because in what is next, we will need to
know more about feature computers, despite dealing with feature specs.
A feature computer service is what link specs to computer, and they
know expose this information.