Make Multiplicity a field of Feature.
It makes no sense to have feature projections each with a possibly
different multiplicity, because they should be coupled.
If there is a value for a ch1 for one projection, there should also
be a value for the same channel for another projection in the same
feature.
Example: mean and std of spot intensity.