ColorMap: Classical linear LUT.
Associate a color to a scalar number. The color is taken
from a linearinterpolation within a list of colors at specific
positions between 0 and 1.A specific color is returned for
Double.NaN values.

A list of colormaps contains defaults plus the colormaps
that are deserialized from a YAML file. The class also sources the
ImageJ LUT files (*.lut) that are stored in the ~/.mastodon/luts/
folder.