Fix slow table interactivity when display the context of large graphs.
To display context objects in the table, we are using Swing table
RowFilter. So the table has the full list of objects and when a context
is activated, the table  filters this list by asking if the object
to display is in the context iterable.
This is slow even with binary search.

This commit works around this in a non subtle way. When a context is
selected we put its content in the table model and only displays it.
We make sure we only display the part of the context that is inside
the full list of objects by taking an intersection when the filter
method is called.