Set enabled flag of selection menu items from EDT thread.
Ok, so this is downstream a bug that caused Mastodon to freeze
completely when using the semi-automatic tracker.
When the semi-automatic tracker is launched from a 'normal' thread,
it completely freezed Mastodon.
The reason is that the tracker changes the selection model when a
spot is created or discovered. The selection model listeners in the
SelectionActions class that toggle whether some menu items (related
to selection) are enabled or not.
Since the listeners were not triggered from the EDT, this caused the
freeze.

To fix this problem, this commit simply ensures that the setEnabled
method from related Action is called from the EDT with
SwingUtilities.involkeLater.