Harness 2D datasets in BDV views.
This commit brings a convenient way to deal with 2D datasets in
Mastodon-app, specifically for the BDV view.
A special BehaviourTransformEventHandler is passed to the viewer
panel when the data is found to be 2D. It features custom behaviours.
Fine tuning is done in the MamutViewBdv: we remove actions that align
the data with XZ or ZY planes, and sets the transform at z=0.

This was adapted from a similar feature in the MaMuT software.