Utility class used to automatically assign sensible mnemonics in JMenus.
One instance of this iterator is instantiated per menu. As the menu is built,
add the menu item with #add(AbstractButton)}method. When all the menu
items have been created and added, call the #assignMnenonics()}methods. The mnemonic assigned will be unique for all the menu items, and
will be taken from the first letter of the name not already assigned to
another mnemonic. If all letters of the name are taken, then no mnemonics are
assigned.