Utility to register number listeners on components.
The listeners of this class are KeyListeners that listen to the
user typing a number, building a string representation as he types,
then converting the string to a double after a wait time. So that
if the user types:

 '4', '9', '.', '3,

then wait for 1 second without typing anything, the component will
by notified with the number 49.3.

Different methods allow to register a listener that can listen
to positive only double or integer numbers.