Fix TrackScheme panel bounds.
Before this commit, some of the TrackScheme panel space was taken by
superfluous ranges. For instance, the top-left corner of the panel
could go up to the time-point -0.5 and to X=-0.5.
The same for the bottom-left (stop half a time-point after the last
time-point) and the other corners.

This commit fixes this problem, both in the screen transform handler
and with the JScrollBar bounds.
Fixes #92