A Translation animator that also linearly interpolates over time-points.
Start and target time-points are to be specified. Because the animator
AbstractTransformAnimator hierarchy does not specify animating time,
this class works with a hook in the shape of an IntConsumer.

The consumer will be called at every step of the animation, and
it is the caller responsibility to ensure that this consumer
sets the time of the object to animate.