Class to manage several visibility modes of a graph.
4 modes are implemented:
- ALL: everything is visible.
- NONE: nothing is visible.
- FOCUS: the track of the focused vertex is visible.
- SELECTION: the current selection is visible.

Each visibility mode is identified by the enum VisibilityMode.
The current visibility mode stored in an instance of this class
can be accessed with getMode() and cycled with nextMode().

Visibilities can return useful implementation of the Visibility
interface, that can state whether a vertex or an edge is visible.
The Visibility instance behavior returned by getVisibility()
matches the visibility mode.

Listeners implementing the VisibilityListener interface can
register to an instance of this class, to be notified when the
visibility of the current mode changes. This might happens for
instance:
- when the visibility mode changes;
- when the selection changes in the VisibilityMode.SELECTION mode;
- when the focus changes in the VisibilityMode.TRACK_OF_FOCUSED_VERTEX
mode.