Use Visibilities in OverlayGraphRenderer.
The Visibility objects are used every-time one has to decide on the
visibility of a vertex or an edge. This includes calls to methods
like #getVertexAt() or #getEdgeAt(). Also the methods called by
the BdvContextProvider, to keep visibility consistent.

The visibility itself is not super clever. We simply rely on the
isVisible() method inside for loops over the vertices and edges
returned by the clip convex polytope.