Exploit the sources-to-compute settings in gauss mean intensity.
More complicated than it sounds, because we also use the update stack
for this feature computer.

- If we compute the feature for ch1, then active ch2, there is no way
built-in to signal that since the feature was not calculated for ch2,
all spots should be parsed again for this channel.

- Rather than rewrite the update stack mechanism, that would also
require revisiting the feature serialization, I relied on determining
'manually' if the feature is present in a channel by directly checking
if the feature map is empty for this channel.