WIP: Copy TrackScheme view to future grapher view.
We want to have a grapher view not based on JFreeChat, but adapting
what Tobias did for TrackScheme. Something well integrated with other
views, in sync with otheer views and possibly with LoD.
Right now this is just an almost litteral copy of the TrackScheme
package, that will servee now as a basis for remaking it for X & Y
plots of the feature data.