Fix a bug preventing to assign the 10th tag of a tag-set.
The action that lets you choose the tag in TrackScheme with an overlay
(default key: Y in TrackScheme) did not work perfectly when a tag-set
have more than 10 tags.

Normally you should be able to select the tag by pressing its number on
the keyboard. It works also when the number of tags is larger than 10,
with a special listener that can wait for several key-presses before
generating an event.

But it did not work for the number 10. The reason was that another
action was assigned to the key '0', that was treated separately, and
messed with the key-presses.

This commit fixes this by handling the '0' action with the same number
listener when we have more than 10 tags.
Noted bu Arianne Bercowsky.